requirements.txt
biopython
jupyter
matplotlib
